Add third example that creates grid of cubes with 2 different pipelines
Every other entity has a different pipeline assigned to it. The idea is to
try and get a baseline for performance when constantly switching between
different pipelines and rendering states with many different objects.

The objects are also all having their transforms updated every frame,
which will be another source of CPU side slowness.

Results so far were frametimes around 90-100ms on my laptop. Excited to
see that number drop.