Bin render objects by pipelineId
Now RenderPasses track their renderObjects using an AoA, where each
element is a list of RenderObjects with the same PipelineId. The idea is
that we reduce the number of state changes per draw call to just
InputAssembler stuff (vertex and index data) and shader resource binding

On my desktop, this saves about 10ms for 10,000 draw calls, and about
3-4ms for 2,500 draw calls. It's not mind blowing, but the bottleneck
is now purely the calls to DrawIndexed().